requirements.txt
apscheduler
requests
git+https://github.com/pyrogram/pyrogram@asyncio-dev
TgCrypto
pydrive
sqlalchemy
psycopg2
bs4
pyDownload
pillow
gtts
gitpython
pafy
selenium
youtube_dl
urbandict
googletrans
currencyconverter
gsearch
pySmartDL
telegraph
coffeehouse
